requirements.txt
requests
python-dotenv
pyyaml
pathlib
markdown
tabulate
jinja2
